+++ tests/e2e/cases/url_object_methods/in.ts
@@ -1,10 +1,17 @@
 const u = new URL("https://example.com:8443/a/b?x=1#frag");
+let seen = "";
+
+function mark(label: string): string {
+  seen += label;
+  return label;
+}
 
 console.log("toString:", u.toString());
 console.log("locale:", u.toLocaleString());
 console.log("json:", u.toJSON());
 console.log("concat:", "url=" + u);
 console.log("value:", u.valueOf().pathname, u.valueOf().search);
+console.log("ignored:", u.toString(mark("s")), u.toLocaleString(mark("l")), u.toJSON(mark("j")), u.valueOf(mark("v")) === u, seen);
 console.log("own:", Object.keys(u).length, Object.getOwnPropertyNames(u).length, Object.hasOwn(u, "href"));
 const descs: any = Object.getOwnPropertyDescriptors(u);
 console.log("desc:", String(Object.getOwnPropertyDescriptor(u, "href")), Object.keys(descs).length);